requirements.txt
huggingface_hub[cli]
transformers
torch
python-dotenv
attrs
accelerate
pandas
seaborn
matplotlib
datasets
trl